+++ web/app/controllers/page/edit.js
@@ -1,13 +1,13 @@
 import Ember from 'ember';
+import { computedAutosave } from 'ember-autosave';
 
 export default Ember.Controller.extend({
-  actions: {
-    discardChanges: function () {
-      this.get('model').rollback();
-      this.transitionToRoute('page.show', this.get('model'));
-    },
+  page: computedAutosave('model'),
 
-    saveChanges: function () {
+  saveDisallowed: Ember.computed.not('page.isDirty'),
+
+  actions: {
+    finishEditing: function () {
       this.get('model').save();
       this.transitionToRoute('page.show', this.get('model'));
     },